+++ src/conversation/menu.py
@@ -8,7 +8,6 @@ from core.config import TRELLO_BORD_ID, URL_SERVICE_RULES
 from core.logger import logger
 from decorators.logger import async_error_logger
 from service.api_client import APIService
-from service.repeat_message import repeat_message_after_1_hour_callback
 
 
 @async_error_logger(name="conversation.menu_commands.menu", logger=logger)
@@ -54,27 +53,6 @@ async def button_reaction_callback(update: Update, context: ContextTypes.DEFAULT
     return states.MENU_STATE
 
 
-async def done_bill_callback_handler(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """
-    Delete job from JobQueue
-    """
-    query = update.callback_query
-    user_id = query.from_user.id
-    current_jobs = context.job_queue.get_jobs_by_name(f"send_{user_id}_bill_until_complete")
-    for job in current_jobs:
-        job.schedule_removal()
-    await query.edit_message_text(text="Не будем напоминать до следующего месяца")
-    await query.answer()  # close progress bar in chat
-
-
-async def skip_bill_callback_handler(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Delete button under message"""
-    query = update.callback_query
-    data = query.message
-    await query.edit_message_text(text=data.text_markdown_v2_urled)
-    await query.answer()  # close progress bar in chat
-
-
 @async_error_logger(name="conversation.requests.button_statistic_month_callback", logger=logger)
 async def button_statistic_month_callback(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
     """
@@ -108,9 +86,6 @@ menu_conversation = ConversationHandler(
             CallbackQueryHandler(button_reaction_callback, pattern=callback_data.CALLBACK_ACTUAL_REQUESTS_COMMAND),
             CallbackQueryHandler(button_reaction_callback, pattern=callback_data.CALLBACK_OVERDUE_REQUESTS_COMMAND),
             CallbackQueryHandler(configurate_timezone, pattern=callback_data.CALLBACK_CONFIGURATE_TIMEZONE_COMMAND),
-            CallbackQueryHandler(repeat_message_after_1_hour_callback, pattern=callback_data.CALLBACK_REPEAT_COMMAND),
-            CallbackQueryHandler(done_bill_callback_handler, pattern=callback_data.CALLBACK_DONE_BILL_COMMAND),
-            CallbackQueryHandler(skip_bill_callback_handler, pattern=callback_data.CALLBACK_SKIP_BILL_COMMAND),
         ],
         **states_timezone_conversation_dict,
     },
